Refactored and simplified HdxPartitionReader to use java Stream;

+++ src/main/scala/io/hydrolix/connectors/partitionreader/HdxPartitionReader.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Hydrolix Inc.
+ * Copyright (c) 2023-2024 Hydrolix Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,8 +19,7 @@ package io.hydrolix.connectors.partitionreader
 import java.io._
 import java.nio.file.Files
 import java.util.Base64
-import java.util.concurrent.ArrayBlockingQueue
-import java.util.concurrent.atomic.AtomicLong
+import java.util.concurrent.{ArrayBlockingQueue, CountDownLatch}
 import java.util.zip.GZIPInputStream
 import scala.collection.mutable
 import scala.sys.process.{Process, ProcessIO}
@@ -37,12 +36,16 @@ import io.hydrolix.connectors.{Etc, HdxConnectionInfo, HdxPartitionScanPlan, Hdx
 object HdxPartitionReader {
   private val log = LoggerFactory.getLogger(getClass)
 
+  private val dontDelete = System.getenv("hdx_reader_no_delete") != null
+
   /**
    * A private parent directory for all temp files belonging to a single instance
    */
   private lazy val hdxReaderTmp = {
     Files.createTempDirectory("hdx_reader").also { path =>
-      new RmRfThread(path.toFile).hook()
+      if (!dontDelete) {
+        new RmRfThread(path.toFile).hook()
+      }
     }.toFile
   }
 
@@ -111,76 +114,49 @@ object HdxPartitionReader {
   private val StderrFilterR = """^read hdx_partition=(.*?) rows=(.*?) values=(.*?) in (.*?)$""".r
   private val DockerPathPrefix = "/hdx-reader"
   private val HdxFs = "hdxfs"
-}
 
-/**
- * TODO:
- *  - Allow secrets to be retrieved from secret services, not just config parameters
- */
-trait HdxPartitionReader[T <: AnyRef] {
-  private val log = LoggerFactory.getLogger(getClass)
+  private def setupTempFiles(info: HdxConnectionInfo, storage: HdxStorageSettings): (File, Option[File]) = {
+    val turbineIniBefore = TurbineIni(storage, info.cloudCred1, info.cloudCred2, if (info.turbineCmdDockerName.isDefined) s"$DockerPathPrefix/$HdxFs" else hdxFsTmp.getAbsolutePath)
 
-  val info: HdxConnectionInfo
-  val storage: HdxStorageSettings
-  val primaryKeyName: String
-  val scan: HdxPartitionScanPlan
+    lazy val (turbineIniAfter, credsTempFile) = if (storage.cloud == "gcp" || storage.cloud == "gcs") {
+      val gcsKeyFile = File.createTempFile("turbine_gcs_key", ".json", hdxReaderTmp)
 
-  protected val doneSignal: T
-  protected def outputFormat: String
-  protected def handleStdout(stdout: InputStream): Unit
+      val turbineIni = Using.Manager { use =>
+        // For gcs, cloudCred1 is a base64(gzip(gcs_service_account_key.json)) and cloudCred2 is unused
+        val gcsKeyB64 = Base64.getDecoder.decode(info.cloudCred1)
 
-  import HdxPartitionReader._
+        val gcsKeyBytes = ByteStreams.toByteArray(use(new GZIPInputStream(new ByteArrayInputStream(gcsKeyB64))))
+        use(new FileOutputStream(gcsKeyFile)).write(gcsKeyBytes)
 
-  protected val stdoutQueue = new ArrayBlockingQueue[T](1024)
-  // Cache because PartitionReader says get() should always return the same record if called multiple times per next()
-  @volatile private var data: T = _
-  @volatile private var recordsEmitted = 0
-  protected val expectedLines = new AtomicLong(1) // 1, not 0, to make room for doneSignal
-  private val stderrLines = mutable.ListBuffer[String]()
+        val gcsKeyPath = if (info.turbineCmdDockerName.isDefined) {
+          s"$DockerPathPrefix/${gcsKeyFile.getName}"
+        } else {
+          gcsKeyFile.getAbsolutePath
+        }
+        val turbineIniWithGcsCredsPath = turbineIniBefore.replace("%CREDS_FILE%", gcsKeyPath)
 
-  // This is done early so it can crash before creating temp files etc.
-  private val schema = scan.schema.fields.map { fld =>
-    HdxOutputColumn(fld.name, scan.hdxCols.getOrElse(fld.name, sys.error(s"No HdxColumnInfo for ${fld.name}")).hdxType)
+        turbineIniWithGcsCredsPath
+      }.get
+
+      (turbineIni, Some(gcsKeyFile))
+    } else {
+      // AWS doesn't need any further munging of turbine.ini
+      (turbineIniBefore, None)
+    }
+
+    // TODO don't create a duplicate file per partition, use a content hash or something
+    lazy val turbineIniTmp = File.createTempFile("turbine", ".ini", hdxReaderTmp)
+    resource(new FileOutputStream(turbineIniTmp)) {
+      _.write(turbineIniAfter.getBytes("UTF-8"))
+    }
+
+    (turbineIniTmp, credsTempFile)
   }
 
-  private val turbineIniBefore = TurbineIni(storage, info.cloudCred1, info.cloudCred2, if (info.turbineCmdDockerName.isDefined) s"$DockerPathPrefix/$HdxFs" else hdxFsTmp.getAbsolutePath)
-
-  private lazy val (turbineIniAfter, credsTempFile) = if (storage.cloud == "gcp" || storage.cloud == "gcs") {
-    val gcsKeyFile = File.createTempFile("turbine_gcs_key", ".json", hdxReaderTmp)
-
-    val turbineIni = Using.Manager { use =>
-      // For gcs, cloudCred1 is a base64(gzip(gcs_service_account_key.json)) and cloudCred2 is unused
-      val gcsKeyB64 = Base64.getDecoder.decode(info.cloudCred1)
-
-      val gcsKeyBytes = ByteStreams.toByteArray(use(new GZIPInputStream(new ByteArrayInputStream(gcsKeyB64))))
-      use(new FileOutputStream(gcsKeyFile)).write(gcsKeyBytes)
-
-      val gcsKeyPath = if (info.turbineCmdDockerName.isDefined) {
-        s"$DockerPathPrefix/${gcsKeyFile.getName}"
-      } else {
-        gcsKeyFile.getAbsolutePath
-      }
-      val turbineIniWithGcsCredsPath = turbineIniBefore.replace("%CREDS_FILE%", gcsKeyPath)
-
-      turbineIniWithGcsCredsPath
-    }.get
-
-    (turbineIni, Some(gcsKeyFile))
-  } else {
-    // AWS doesn't need any further munging of turbine.ini
-    (turbineIniBefore, None)
-  }
-
-  // TODO don't create a duplicate file per partition, use a content hash or something
-  private lazy val turbineIniTmp = File.createTempFile("turbine", ".ini", hdxReaderTmp)
-  resource(new FileOutputStream(turbineIniTmp)) {
-    _.write(turbineIniAfter.getBytes("UTF-8"))
-  }
-
-  // TODO does anything need to be quoted here?
-  //  Note, this relies on a bunch of changes in hdx_reader that may not have been merged to turbine/turbine-core yet,
-  //  see https://hydrolix.atlassian.net/browse/HDX-3779
-  private val hdxReaderProcess = {
+  private def prepareCmdAndArgs(dockerName: Option[String], turbineIniTmp: File, outputFormat: String, scan: HdxPartitionScanPlan, primaryKeyName: String, schema: List[HdxOutputColumn]): List[String] = {
+    // TODO does anything need to be quoted here?
+    //  Note, this relies on a bunch of changes in hdx_reader that may not have been merged to turbine/turbine-core yet,
+    //  see https://hydrolix.atlassian.net/browse/HDX-3779
     val schemaStr = JSON.objectMapper.writeValueAsString(schema)
 
     val exprArgs = {
@@ -196,7 +172,7 @@ trait HdxPartitionReader[T <: AnyRef] {
       }
     }
 
-    val turbineIniPath = if (info.turbineCmdDockerName.isDefined) {
+    val turbineIniPath = if (dockerName.isDefined) {
       s"$DockerPathPrefix/${turbineIniTmp.getName}"
     } else {
       turbineIniTmp.getAbsolutePath
@@ -211,29 +187,30 @@ trait HdxPartitionReader[T <: AnyRef] {
       "--schema", schemaStr
     ) ++ exprArgs
 
-    val (cmd, args) = info.turbineCmdDockerName match {
+    dockerName match {
       case Some(imageName) =>
         // docker run -v ~/dev/hydrolix/hdx-spark/src/main/resources/linux-x86-64:/hdx-spark -w /hdx-spark ubuntu:22.04 ./turbine_cmd -a hdx_reader ...
-        (
+        List(
           "docker", // TODO this assumes docker is on the PATH
-          List(
-            "run",
-            "-a", "STDOUT",
-            "-a", "STDERR",
-            "-v", s"${hdxReaderTmp.getAbsolutePath}:$DockerPathPrefix",
-            imageName,
-            s"$DockerPathPrefix/${turbineCmdTmp.getName}"
-          ) ++ turbineCmdArgs
-        )
+          "run",
+          "-a", "STDOUT",
+          "-a", "STDERR",
+          "-v", s"${hdxReaderTmp.getAbsolutePath}:$DockerPathPrefix",
+          imageName,
+          s"$DockerPathPrefix/${turbineCmdTmp.getName}"
+        ) ++ turbineCmdArgs
       case None =>
-        (turbineCmdTmp.getAbsolutePath, turbineCmdArgs)
+        turbineCmdTmp.getAbsolutePath +: turbineCmdArgs
     }
+  }
 
-    log.info(s"Running $cmd ${args.mkString(" ")}")
-
+  private def launch(cmdAndArgs: List[String],
+                   handleStdout: InputStream => Unit,
+                    stderrLines: mutable.ListBuffer[String])
+                               : Process =
+  {
     Process(
-      cmd,
-      args
+      cmdAndArgs
     ).run(
       new ProcessIO(
         {
@@ -252,42 +229,91 @@ trait HdxPartitionReader[T <: AnyRef] {
       )
     )
   }
+}
 
-  def next(): Boolean = {
-    val got = try {
-      stdoutQueue.take() // It's OK to block here, we'll always have doneSignal...
-    } catch {
-      case _: InterruptedException =>
-        // ...But if we got killed while waiting, don't be noisy about it
-        Thread.currentThread().interrupt()
-        return false
+/**
+ * Grab `stream` to consume data. It blocks the calling thread!
+ *
+ * TODO:
+ *  - Allow secrets to be retrieved from secret services, not just config parameters
+ */
+abstract class HdxPartitionReader[T >: Null <: AnyRef](doneSignal: T, outputFormat: String) {
+  private val log = LoggerFactory.getLogger(getClass)
+
+  val info: HdxConnectionInfo
+  val storage: HdxStorageSettings
+  val primaryKeyName: String
+  val scan: HdxPartitionScanPlan
+
+  /**
+   * Called in a new thread to handle stdout from the hdx_reader child process. Call `enqueue` with a new value or
+   * `doneSignal`. Implementer should close the stream before calling `enqueue(doneSignal)`.
+   */
+  protected def handleStdout(stdout: InputStream): Unit
+
+  /** Blocks until the first enqueue call, whether a value or `doneSignal` */
+  private val started = new CountDownLatch(1)
+  /** Blocks until after enqueue(doneSignal) */
+  private val finished = new CountDownLatch(1)
+  private val stdoutQueue = new ArrayBlockingQueue[T](1024)
+
+  /**
+   * Called by the `handleStdout` function to enqueue a value, or doneSignal
+   */
+  def enqueue(value: T): Unit = {
+    started.countDown()
+    if (value eq doneSignal) {
+      finished.countDown()
     }
+    stdoutQueue.put(value) // TODO it's possible this could block forever if the queue is full and the consumer is stalled
+  }
 
-    if (got eq doneSignal) {
+  // This is done early so it can crash before creating temp files etc.
+  private val schema = scan.schema.fields.map { fld =>
+    HdxOutputColumn(fld.name, scan.hdxCols.getOrElse(fld.name, sys.error(s"No HdxColumnInfo for ${fld.name}")).hdxType)
+  }
+
+  import HdxPartitionReader._
+
+  private val (turbineIniTmp, credsTempFile) = setupTempFiles(info, storage)
+
+  private val cmdAndArgs = prepareCmdAndArgs(info.turbineCmdDockerName, turbineIniTmp, outputFormat, scan, primaryKeyName, schema)
+  private val stderrLines = mutable.ListBuffer[String]()
+
+  log.info(s"Running ${cmdAndArgs.mkString(" ")}")
+
+  private val hdxReaderProcess = launch(cmdAndArgs, handleStdout, stderrLines)
+
+  val stream = {
+    // TODO this should have a timeout just in case the implementor never enqueues anything
+    started.await()
+
+    stdoutQueue.stream().takeWhile(_ ne doneSignal).onClose { () =>
+      finished.await()
+
       // There are definitely no more records, stdout is closed, now we can wait for the sweet release of death
       val exit = hdxReaderProcess.exitValue()
       val err = stderrLines.mkString("\n  ", "\n  ", "\n")
+
       if (exit != 0) {
-        throw new RuntimeException(s"turbine_cmd process exited with code $exit; stderr was: $err")
+        sys.error(s"turbine_cmd process exited with code $exit; stderr was $err")
       } else {
         if (err.trim.nonEmpty) log.warn(s"turbine_cmd process exited with code $exit but stderr was: $err")
-        false
       }
-    } else {
-      this.data = got
-
-      recordsEmitted += 1
-
-      expectedLines.decrementAndGet() > 0
     }
   }
 
-  def get(): T = data
-
   def close(): Unit = {
-    log.info(s"$recordsEmitted records read from ${scan.partitionPath}")
-    Try(turbineIniTmp.delete())
-    credsTempFile.foreach(f => Try(f.delete()))
+    try {
+      if (hdxReaderProcess.isAlive()) {
+        log.info("Child process is still alive; killing")
+        hdxReaderProcess.destroy()
+      }
+    } finally {
+      if (!dontDelete) {
+        Try(turbineIniTmp.delete())
+        credsTempFile.foreach(f => Try(f.delete()))
+      }
+    }
   }
 }
-
